fix(web): enforce getTenant server/client boundary so next build compiles (CTO-259)
`lib/getTenant.ts` is server-only (it lazily loads `@clerk/nextjs/server`, which
pulls in `server-only`), but three client-imported libs re-exported it through
pure helpers, dragging Clerk's server bundle into the client graph and failing
`next build` with "'server-only' cannot be imported from a Client Component".

The offending chains the build revealed:

  app/connectors/ConnectorTable.tsx  -> lib/costConnectors.ts  -> lib/getTenant.ts
  app/connectors/RevenueUpload.tsx   -> lib/revenueUpload.ts   -> lib/getTenant.ts
  app/settings/budgets/BudgetManager -> lib/budgets.ts         -> lib/getTenant.ts

Fix:
- Guard `getTenant.ts` with `import "server-only"` so any client import fails
  loudly at the source, and move its client-safe symbols (`ORG_ADMIN_ROLE`,
  `ResolvedTenant`) into `lib/tenantShared.ts`, re-exported for server callers.
- Split each offending lib: the client-safe constants, types and pure helpers
  move to `*Shared.ts` (costConnectorsShared, budgetsShared, revenueUploadShared),
  re-exported from the server lib so server call sites are unchanged; the three
  client components now import from the shared modules.
- Alias `server-only` to its empty build in vitest so Node-run tests can import
  server modules under test, matching Next's `react-server` export condition.

The full build also prerenders Clerk pages; make the keyless build (the dev
escape hatch, §10, how CI and `make up` run without Clerk keys) succeed:
- Dev-gate `/settings/members` (it is nav-reachable) and mark the Clerk auth
  pages (`select-org`, `sign-in`, `sign-up`) `force-dynamic` so the keyless
  build does not statically prerender a Clerk component that needs a
  publishableKey.
- Run the web CI Build step with `TALLY_DEV_TENANT` set, the same keyless mode
  the vitest suite and `make up` already use.

Verified: typecheck, lint, test (only the documented ClickHouse-reachability
case fails locally), and build all green.
