+++ web/app/sign-in/[[...sign-in]]/page.tsx
@@ -3,6 +3,10 @@
 // in an organization.
 import { SignIn } from "@clerk/nextjs";
 
+// Rendered per request, not at build: the keyless build (dev escape hatch / CI) has no Clerk
+// publishableKey, so this Clerk component must resolve at request time on the product path (CTO-259).
+export const dynamic = "force-dynamic";
+
 export default function SignInPage() {
   return (
     <div className="flex min-h-[70vh] items-center justify-center">
