+++ web/app/sign-up/[[...sign-up]]/page.tsx
@@ -2,6 +2,10 @@
 // Clerk-hosted sign-up (Initiative 1, §7). Public route.
 import { SignUp } from "@clerk/nextjs";
 
+// Rendered per request, not at build: the keyless build (dev escape hatch / CI) has no Clerk
+// publishableKey, so this Clerk component must resolve at request time on the product path (CTO-259).
+export const dynamic = "force-dynamic";
+
 export default function SignUpPage() {
   return (
     <div className="flex min-h-[70vh] items-center justify-center">
